refactor(commands): move resolve_addr into commands::utils
`inspect path` reached into a sibling subcommand (`super::revdeps::resolve_addr`)
for target-argument resolution, which reads as if the two commands were
related. The helper is generic CLI argument handling, so it belongs beside
`resolve_matcher` in `commands::utils`; both call sites now import it from
there. Its tests move with it.
