+++ src/commands/inspect/revdeps.rs
@@ -1,4 +1,3 @@
-use std::path::Path;
 use std::sync::Arc;
 
 use anyhow::Context;
@@ -9,8 +8,9 @@ use futures::StreamExt;
 use crate::commands::GlobalOptions;
 use crate::commands::bootstrap;
 use crate::commands::completion::complete_target_addr;
+use crate::commands::utils::resolve_addr;
 use crate::engine::Engine;
-use crate::htaddr::{self, Addr};
+use crate::htaddr::Addr;
 use crate::htmatcher::Matcher;
 use crate::htpkg::{self, PkgBuf};
 use crate::tui::{self, App, AppContext, BufferedStdout, LogSink};
@@ -83,40 +83,6 @@ pub fn execute(args: &Args, sink: LogSink, global: &GlobalOptions) -> anyhow::Re
     bootstrap::block_on(execute_async(args.clone(), sink, global.clone()))?
 }
 
-/// Resolve a CLI target argument into an `Addr`, relative to package `cwp` under
-/// workspace `root`.
-///
-/// First tries to parse `input` as a (possibly relative) target address —
-/// `//pkg:name`, `:name`, `./pkg:name`. If that fails, `input` is treated as a
-/// path: when it points at an existing file, the file's `fs` target
-/// (`//@heph/fs:file@f=<root-relative path>`) is used; otherwise the original
-/// parse error is surfaced.
-fn resolve_addr_in(input: &str, cwp: &PkgBuf, root: &Path) -> anyhow::Result<Addr> {
-    match htaddr::parse_addr_with_base(input, cwp) {
-        Ok(addr) => Ok(addr),
-        Err(parse_err) => {
-            // Not a valid address — fall back to the file-path sugar, but only
-            // when it actually names a file on disk. Otherwise the address parse
-            // error is the useful one to show.
-            if let Ok(rel) = htpkg::join_rel_checked(cwp.as_str(), input)
-                && root.join(&rel).is_file()
-            {
-                return Ok(crate::pluginfs::file_addr(&rel));
-            }
-            Err(parse_err).with_context(|| format!("parse {input}"))
-        }
-    }
-}
-
-/// `resolve_addr_in` against the current working package and workspace root.
-pub(super) fn resolve_addr(input: &str) -> anyhow::Result<Addr> {
-    resolve_addr_in(
-        input,
-        &crate::engine::get_cwp()?,
-        &crate::engine::get_root()?,
-    )
-}
-
 async fn execute_async(args: Args, sink: LogSink, global: GlobalOptions) -> anyhow::Result<()> {
     let addr = resolve_addr(args.addr.as_ref())?;
     let scope = match &args.scope {
@@ -133,78 +99,4 @@ async fn execute_async(args: Args, sink: LogSink, global: GlobalOptions) -> anyh
     };
     let interactive = tui::should_use_tui(global.no_tui);
     tui::run_app(app, sink, interactive, shutdown).await
-}
-
-#[cfg(test)]
-mod tests {
-    use super::resolve_addr_in;
-    use crate::htaddr::parse_addr;
-    use crate::htpkg::PkgBuf;
-
-    /// A tempdir root with `rel` touched as an empty file.
-    fn root_with_file(rel: &str) -> tempfile::TempDir {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let path = dir.path().join(rel);
-        std::fs::create_dir_all(path.parent().unwrap()).unwrap();
-        std::fs::write(&path, "").unwrap();
-        dir
-    }
-
-    #[test]
-    fn existing_bare_file_resolves_to_fs_file_addr() {
-        // A bare path is not a valid address, so it falls back to the file
-        // sugar — the fs file target keyed by the root-relative path.
-        let root = root_with_file("cmd/server/data.txt");
-        let addr = resolve_addr_in("data.txt", &PkgBuf::from("cmd/server"), root.path()).unwrap();
-        assert_eq!(addr, crate::pluginfs::file_addr("cmd/server/data.txt"));
-    }
-
-    #[test]
-    fn bare_subdir_file_resolves_against_package() {
-        let root = root_with_file("cmd/server/src/main.rs");
-        let addr =
-            resolve_addr_in("src/main.rs", &PkgBuf::from("cmd/server"), root.path()).unwrap();
-        assert_eq!(addr, crate::pluginfs::file_addr("cmd/server/src/main.rs"));
-    }
-
-    #[test]
-    fn unparseable_missing_path_surfaces_parse_error() {
-        // Not an address and not a file on disk → the address parse error wins.
-        let dir = tempfile::tempdir().unwrap();
-        let err = resolve_addr_in("data.txt", &PkgBuf::from("cmd/server"), dir.path()).unwrap_err();
-        assert!(format!("{err:#}").contains("parse data.txt"), "{err:#}");
-    }
-
-    #[test]
-    fn dot_slash_path_to_existing_file_uses_fs_target() {
-        // `./somefile.txt` is not a valid address (a relative path ref must name
-        // a target), so an existing file takes the fs file sugar.
-        let root = root_with_file("cmd/server/somefile.txt");
-        let addr =
-            resolve_addr_in("./somefile.txt", &PkgBuf::from("cmd/server"), root.path()).unwrap();
-        assert_eq!(addr, crate::pluginfs::file_addr("cmd/server/somefile.txt"));
-    }
-
-    #[test]
-    fn dot_slash_relative_target_with_explicit_name() {
-        // The explicit `:name` form is a target address, parsed before any disk
-        // check.
-        let dir = tempfile::tempdir().unwrap();
-        let addr = resolve_addr_in("./sub:thing", &PkgBuf::from("cmd/server"), dir.path()).unwrap();
-        assert_eq!(addr, parse_addr("//cmd/server/sub:thing").unwrap());
-    }
-
-    #[test]
-    fn plain_addr_is_parsed_verbatim() {
-        let dir = tempfile::tempdir().unwrap();
-        let addr = resolve_addr_in("//lib:core", &PkgBuf::from("cmd/server"), dir.path()).unwrap();
-        assert_eq!(addr, parse_addr("//lib:core").unwrap());
-    }
-
-    #[test]
-    fn colon_relative_target_resolves_against_package() {
-        let dir = tempfile::tempdir().unwrap();
-        let addr = resolve_addr_in(":mytarget", &PkgBuf::from("cmd/server"), dir.path()).unwrap();
-        assert_eq!(addr, parse_addr("//cmd/server:mytarget").unwrap());
-    }
 }
