+++ src/commands/utils.rs
@@ -1,7 +1,40 @@
+use std::path::Path;
+
+use crate::htaddr::Addr;
 use crate::htmatcher::Matcher;
 use crate::htpkg::PkgBuf;
 use crate::{engine, htaddr, htpkg, htquery};
 use anyhow::Context;
+
+/// Resolve a CLI target argument into an `Addr`, relative to package `cwp` under
+/// workspace `root`.
+///
+/// First tries to parse `input` as a (possibly relative) target address —
+/// `//pkg:name`, `:name`, `./pkg:name`. If that fails, `input` is treated as a
+/// path: when it points at an existing file, the file's `fs` target
+/// (`//@heph/fs:file@f=<root-relative path>`) is used; otherwise the original
+/// parse error is surfaced.
+fn resolve_addr_in(input: &str, cwp: &PkgBuf, root: &Path) -> anyhow::Result<Addr> {
+    match htaddr::parse_addr_with_base(input, cwp) {
+        Ok(addr) => Ok(addr),
+        Err(parse_err) => {
+            // Not a valid address — fall back to the file-path sugar, but only
+            // when it actually names a file on disk. Otherwise the address parse
+            // error is the useful one to show.
+            if let Ok(rel) = htpkg::join_rel_checked(cwp.as_str(), input)
+                && root.join(&rel).is_file()
+            {
+                return Ok(crate::pluginfs::file_addr(&rel));
+            }
+            Err(parse_err).with_context(|| format!("parse {input}"))
+        }
+    }
+}
+
+/// `resolve_addr_in` against the current working package and workspace root.
+pub fn resolve_addr(input: &str) -> anyhow::Result<Addr> {
+    resolve_addr_in(input, &engine::get_cwp()?, &engine::get_root()?)
+}
 
 /// Resolve the target selection for the `run`/`query` commands. Exactly one of
 /// the query form (`-e '<expr>'`) or the positional form (`<addr>` /
@@ -92,7 +125,75 @@ fn count_matcher(m: &Matcher, c: &mut htelemetry::telemetry::QueryExprCounts) {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::htaddr::parse_addr;
     use crate::htmatcher::Matcher;
+
+    /// A tempdir root with `rel` touched as an empty file.
+    fn root_with_file(rel: &str) -> tempfile::TempDir {
+        let dir = tempfile::tempdir().expect("tempdir");
+        let path = dir.path().join(rel);
+        std::fs::create_dir_all(path.parent().unwrap()).unwrap();
+        std::fs::write(&path, "").unwrap();
+        dir
+    }
+
+    #[test]
+    fn existing_bare_file_resolves_to_fs_file_addr() {
+        // A bare path is not a valid address, so it falls back to the file
+        // sugar — the fs file target keyed by the root-relative path.
+        let root = root_with_file("cmd/server/data.txt");
+        let addr = resolve_addr_in("data.txt", &PkgBuf::from("cmd/server"), root.path()).unwrap();
+        assert_eq!(addr, crate::pluginfs::file_addr("cmd/server/data.txt"));
+    }
+
+    #[test]
+    fn bare_subdir_file_resolves_against_package() {
+        let root = root_with_file("cmd/server/src/main.rs");
+        let addr =
+            resolve_addr_in("src/main.rs", &PkgBuf::from("cmd/server"), root.path()).unwrap();
+        assert_eq!(addr, crate::pluginfs::file_addr("cmd/server/src/main.rs"));
+    }
+
+    #[test]
+    fn unparseable_missing_path_surfaces_parse_error() {
+        // Not an address and not a file on disk → the address parse error wins.
+        let dir = tempfile::tempdir().unwrap();
+        let err = resolve_addr_in("data.txt", &PkgBuf::from("cmd/server"), dir.path()).unwrap_err();
+        assert!(format!("{err:#}").contains("parse data.txt"), "{err:#}");
+    }
+
+    #[test]
+    fn dot_slash_path_to_existing_file_uses_fs_target() {
+        // `./somefile.txt` is not a valid address (a relative path ref must name
+        // a target), so an existing file takes the fs file sugar.
+        let root = root_with_file("cmd/server/somefile.txt");
+        let addr =
+            resolve_addr_in("./somefile.txt", &PkgBuf::from("cmd/server"), root.path()).unwrap();
+        assert_eq!(addr, crate::pluginfs::file_addr("cmd/server/somefile.txt"));
+    }
+
+    #[test]
+    fn dot_slash_relative_target_with_explicit_name() {
+        // The explicit `:name` form is a target address, parsed before any disk
+        // check.
+        let dir = tempfile::tempdir().unwrap();
+        let addr = resolve_addr_in("./sub:thing", &PkgBuf::from("cmd/server"), dir.path()).unwrap();
+        assert_eq!(addr, parse_addr("//cmd/server/sub:thing").unwrap());
+    }
+
+    #[test]
+    fn plain_addr_is_parsed_verbatim() {
+        let dir = tempfile::tempdir().unwrap();
+        let addr = resolve_addr_in("//lib:core", &PkgBuf::from("cmd/server"), dir.path()).unwrap();
+        assert_eq!(addr, parse_addr("//lib:core").unwrap());
+    }
+
+    #[test]
+    fn colon_relative_target_resolves_against_package() {
+        let dir = tempfile::tempdir().unwrap();
+        let addr = resolve_addr_in(":mytarget", &PkgBuf::from("cmd/server"), dir.path()).unwrap();
+        assert_eq!(addr, parse_addr("//cmd/server:mytarget").unwrap());
+    }
 
     #[test]
     fn count_matcher_tallies_node_kinds() {
